AssetFileCreator moved
